client_bot/requirements.txt
# --- Discord Bot ---
discord.py

# --- Cryptography ---
py-cryptodomex

# --- Configuration ---
pyyaml

# --- System Monitoring ---
psutil